backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[all]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
